# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Streaming App API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/superstar/login": {
            "post": {
                "tags": [
                    "Superstar Authentication"
                ],
                "summary": "Superstar Login",
                "description": "Login endpoint for superstars using email or username",
                "operationId": "cd1da0ff025403591120dafab1fdc21f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "description": "Email or username",
                                        "type": "string",
                                        "example": "superstar@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "superstar": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "display_name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "bio": {
                                                    "type": "string",
                                                    "example": "Professional superstar"
                                                },
                                                "price_per_hour": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50
                                                },
                                                "is_available": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "rating": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 4.5
                                                },
                                                "total_followers": {
                                                    "type": "integer",
                                                    "example": 1500
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "active"
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "johndoe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "superstar@example.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "superstar"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-15T10:30:00Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123token456"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Admin access forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Please use admin app for admin access"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "login": [
                                                    "The login field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/superstar/logout": {
            "post": {
                "tags": [
                    "Superstar Authentication"
                ],
                "summary": "Logout superstar",
                "description": "Logout of authenticated superstar and revoke token",
                "operationId": "31de6cff592d1ab209431b66c640cd2c",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/superstar/profile": {
            "put": {
                "tags": [
                    "Superstar Profile"
                ],
                "summary": "Update superstar profile",
                "description": "Update the authenticated superstar's profile information",
                "operationId": "c2c3226452128bf0079a092663f99262",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "display_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Professional superstar"
                                    },
                                    "is_available": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "profile_image": {
                                        "description": "Profile image file (jpeg, png, jpg, gif, max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        },
                                        "superstar": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}